Rename check svg file
